switch to profile fragment
